Wrap pciReadLong / pciWriteLong with macros.
The pciReadLong and pciWriteLong function calls are wrapped with macros for
a couple reasons.  First, this will reduce the "noise" when support for
pci-rework is added to tdfx.  Second, by having them wrapped in macros it is
easier to support both pci-rework and the "classic" interfaces in the same
code base.