Initial pass at porting driver to pci-rework.  DOES NOT WORK.
This is my initial pass at porting xf86-video-tdfx to the pci-rework
interfaces.  For reasons that I'm not able to understand, it doesn't quite
work.  It's clear that some piece of the hardware isn't being initialized
properly, but, by comparing the old code to the new, it's not clear why.

Does not impact the functionality of non-pci-rework builds.